Got encoder working but button won't read. Will document encoder pins in bookstack
