+++ platformio.ini
@@ -88,7 +88,7 @@ lib_deps =
   https://github.com/adafruit/Adafruit_NeoPixel.git
   https://github.com/gianni-carbone/STM32encoder.git
   https://github.com/MajicDesigns/MD_REncoder.git
-  https://github.com/brianlow/Rotary.git
+  https://github.com/CarlosSiles67/Rotary.git
 board_build.ldscript = STM32H750VB_FLASH.ld
 
 [env:H750_dfu]
@@ -106,6 +106,7 @@ debug_speed = 24000
 platform = ststm32
 board = genericSTM32H750VB
 framework = arduino
+upload_protocol = dfu
 build_src_filter = 
   +<*.cpp>
   +<test_encoder.cpp>
@@ -137,4 +138,4 @@ lib_deps =
   https://github.com/adafruit/Adafruit_NeoPixel.git
   https://github.com/gianni-carbone/STM32encoder.git
   https://github.com/MajicDesigns/MD_REncoder.git
-  https://github.com/brianlow/Rotary.git
+  https://github.com/CarlosSiles67/Rotary.git
